remove combine code from async image
