+++ apple/OmnivoreKit/Sources/Views/FeedItem/GridCard.swift
@@ -127,7 +127,7 @@ public struct GridCard: View {
           Spacer()
 
           if let imageURL = item.imageURL {
-            AsyncImage(url: imageURL) { imageStatus in
+            AsyncLoadingImage(url: imageURL) { imageStatus in
               if case let AsyncImageStatus.loaded(image) = imageStatus {
                 image
                   .resizable()
